checkout/endpoint/src/main/java/com/cece/alex/pos/checkout/endpoint/CheckoutController.java
package com.cece.alex.pos.checkout.endpoint;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class CheckoutController {

    private CheckoutService checkoutService;

    @Autowired
    public CheckoutController(CheckoutService checkoutService) {
        this.checkoutService = checkoutService;
    }

    @GetMapping("/orders/{orderId}/total")
    public FinalOrder getTotal(@PathVariable Integer orderId) {
        return checkoutService.getTotal(orderId);
    }
}
